+++ pytest_inmanta/test_parameter/path_parameter.py
@@ -16,9 +16,9 @@
     Contact: code@inmanta.com
 """
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Union
 
-from .parameter import TestParameter
+from pytest_inmanta.test_parameter.parameter import DynamicDefault, TestParameter
 
 
 class PathTestParameter(TestParameter[Path]):
@@ -50,12 +50,13 @@ class PathTestParameter(TestParameter[Path]):
         environment_variable: str,
         usage: str,
         *,
-        default: Optional[Path] = None,
+        default: Optional[Union[Path, DynamicDefault[Path]]] = None,
         key: Optional[str] = None,
         group: Optional[str] = None,
         legacy: Optional["PathTestParameter"] = None,
         is_file: Optional[bool] = None,
         exists: Optional[bool] = None,
+        legacy_environment_variable: Optional[str] = None,
     ) -> None:
         self.is_file = is_file
         self.exists = exists if is_file is None else True
@@ -67,6 +68,7 @@ class PathTestParameter(TestParameter[Path]):
             key=key,
             group=group,
             legacy=legacy,
+            legacy_environment_variable=legacy_environment_variable,
         )
 
     def validate(self, raw_value: object) -> Path: